pyproject.toml
[project]
name = "justkeepswimming"
version = "0.0.0"
description = "A small game about a fish in the deep sea."

readme = "README.md"
license = "GPL-3.0"

requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "psutil>=7.2.1",
    "pygame-ce>=2.5.6",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pycodestyle>=2.14.0",
    "pylint>=4.0.4",
    "types-psutil>=7.2.1.20251231",
]

[tool.black]
line-length = 79

[tool.isort]
profile = "black"

[tool.mypy]
follow_imports = "silent"

[tool.autopep8]
max-line-length = 79

[tool.flake8]
max-line-length = 79

[tool.pylint]
max-line-length = 79

[tool.pycodestyle]
max-line-length = 79


[project.scripts]
client = "justkeepswimming.__main__:main"

[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"
